refactor(markdown): dsl becomes an object
Scala 3 refuses a wildcard export whose prefix is a package —
"Implementation restriction: ... is not a valid prefix for a wildcard
export, as it is a package" — so a facade cannot re-export the 31
combinators and the String-to-Text conversion while `dsl` is a package.
An object can be exported; the move is what makes one door possible.

Source-compatible by construction: `import morphir.langkit.markdown.dsl.*`
and `import ...dsl.given` read identically whether `dsl` names a package
or an object, so no call site changes. Both existing ones — MdWriterTests
and the scalatags WriterFidelityTests — keep their import lines untouched.

DslTests moves up a level with it. A test file declaring `package
morphir.langkit.markdown.dsl` would collide with the object of that name,
and its unqualified `doc`/`h1`/`p` calls were reaching the combinators
through package scope; it now imports them like everyone else.
